binance_trader_pro/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "binance_trader"
version = "0.1.0"
description = "Refactored Binance USDT-M Futures trading system (REST-based, strategies, backtest, live)."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [{name = "GwakJunwoo", email = "lunarsabbth@gmail.com"}]
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "requests>=2.31",
    "websockets>=12.0",
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "ta>=0.11.0"
]

[project.scripts]
binance-trader = "binance_trader.cli:main"


[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["binance_trader*"]
exclude = ["integrations*", "data*", "tests*", "examples*"]


[tool.setuptools.package-data]
"binance_trader" = ["config/*.yaml", "config/.env.example"]
